Changed Labs: - Lab3_AL Removed commented lines that are no longer needed Added an Instance Name to the SQL Role Definition
Added the following roles:
- ChangeSQLServerSettings:
Started the implementation
- InstallNTFSSecurity
Install the NTFSecurity module on every SQLServer
- SetSQLServiceAccountPermissions
Started the implementation

Changed the following roles:
- CreateAlwaysOnAvailabilityGroup
Removed SQLInstance parameter. SQLInstance will be
retrieved over the AL Role properties.
- InstallDBATools
Fixed ComputerName argument
- PrepareSampleDatabaseForAG
Get the SQLInstance of the current VM. If an instance name is configured
the variables SQLInstance and SecondSQLInstance are replaced with the following
format: Computername\InstanceName.
This will be used to find the backup and Restore SQLInstances.
- RestoreSampleDatabase
Removed SQLInstance parameter. SQLInstance will be
retrieved over the AL Role properties.