Added AL Custom Roles: - InstallFoldersAndShares Added this role to add folders and shares without needing a scale out file server.
Changes To AL Custom Roles:
- ChangeSQLServerSettings
Added Parameter FileServerName
Added Logic for Setting the Paths for Backup,Log and Data
Restart all SQL Services so that the changes will take effect
- CreateAlwaysOnAvailabilityGroup
Changed the logic if there is no named instance
- InstallNTFSSecurity
Changed the activity text
- InstallScaleOutFileServerClusterFoldersAndShares
Changed the activity text
- InstallSQLServerCU
Because in testing there was no reliable update process I added the check if all SQL Server Services are running before starting the install. If that was not done the SQLTELEMETRY Service could not be stopped and the update will fail.
Changed some activity texts.
- RestoreSampleDatabase
Added Parameter RestoreFolderPath.
Changed the logic if there is no named instance.
Changed some activity texts.
Changed the restorepath for Restore-DbaDatabase. It is now working without errors.
- SetSQLServiceAccountPermissions
Added parameter ClusterFolderAndShareDefinition
Get the SQL Server Engine Service Account and added the account
to the NTFS permission for the new Data, Log and Backup folders