Changed Custom Roles: - InstallPSResourceGet: Added a while loop to check if PSGallery is available.
Changed Labs:
- Lab3_AL
Removed not needed line
- Lab1_NormalWay
Added the witness information to the secret file
Changed the network configuration from MAC based to static IP Adresses.
Added the PSGallery availability check
Changed the witness step to work with the secret file
Added a configuration to set the browser startup type to automatic
Splitted the Availability Group enabling to run on every node.
Check if SQL Services are running if not start them